School Section

+++ backend/routes/FormRoute.js
@@ -7,8 +7,10 @@ const fs = require('fs');
 const Form = require('../models/Form');
 const Vendor = require('../models/Vendor');
 const IndoorForm = require('../models/IndoorForm');
+const SchoolForm = require('../models/SchoolForm');
 
 const vendorAuth = require('../middleware/vendorAuth');
+const schoolAuth = require('../middleware/schoolAuth');
 
 const awsSES = require('../controller/awsSES');
 
@@ -34,7 +36,8 @@ const fileFilter = (req, file, cb) => {
 	) {
 		cb(null, true);
 	} else {
-		cb(new Error('File type should be png, jpeg or pdf.'), false);
+		req.error = 'File type should be png, jpeg or pdf.';
+		cb(null, false);
 	}
 };
 
@@ -57,13 +60,7 @@ router.post('/add', async (req, res) => {
 			phoneNumber,
 			dateOfVisit,
 			timeOfVisit,
-			temperature,
-			cough,
-			abroadIn14Days,
-			contactIn14Days,
 			noOfPeopleInGroup,
-			vaccinationStatus,
-			covidOver9Months,
 			comments,
 		} = req.body;
 
@@ -74,13 +71,7 @@ router.post('/add', async (req, res) => {
 			phoneNumber,
 			dateOfVisit,
 			timeOfVisit,
-			temperature,
-			cough,
-			abroadIn14Days,
-			contactIn14Days,
 			noOfPeopleInGroup,
-			vaccinationStatus,
-			covidOver9Months: covidOver9Months || false,
 			comments,
 		});
 
@@ -107,14 +98,33 @@ router.post('/add', async (req, res) => {
 // @access  Public
 router.post('/indoor/add', upload.any(), async (req, res, next) => {
 	try {
-		const certificate = req?.files[0];
+		if (req.error) {
+			throw new Error(req.error);
+		}
+		const certificate = req.files[0];
+		const identity = req.files[1];
 
-		const { vendorID, email, fullName, dateOfVisit, timeOfVisit } = req.body;
+		if (!certificate) {
+			throw new Error('Please upload the certificate');
+		}
+		// if (!identity) {
+		// 	await fs.unlinkSync(certificate.path);
+		// 	throw new Error('Please upload your identity.');
+		// }
+		if (certificate && certificate.size / (1024 * 1024) > 5) {
+			throw new Error('File is greater than 5MB. Please select other file.');
+		}
+
+		// if (identity && identity.size / (1024 * 1024) > 5) {
+		// 	throw new Error('File is greater than 5MB. Please select other file.');
+		// }
+
+		const { vendorID, email, fullName, dateOfVisit, timeOfVisit, refNo } =
+			req.body;
 		const vendor = await Vendor.findById(vendorID);
 		if (!vendor || !vendor.vendorName) {
 			throw new Error('The vendor could not be found.');
 		}
-		console.log(vendor.vendorName);
 
 		const formContent = {
 			vendor,
@@ -123,26 +133,23 @@ router.post('/indoor/add', upload.any(), async (req, res, next) => {
 			dateOfVisit,
 			timeOfVisit,
 			certificate,
+			identity,
+			refNo,
 		};
 
-		const newIndoorForm = new IndoorForm({
-			vendorID,
-			email,
-			fullName,
-			dateOfVisit,
-			timeOfVisit,
-		});
-
-		let savedIndoorForm = await newIndoorForm.save();
-		if (!savedIndoorForm) {
-			res.status(400).json({ error: 'Could not save new indoor form data.' });
-		}
 		await awsSES.indorEmailviaAWS_SES(formContent, res);
-		await fs.unlinkSync(certificate.path);
+		setTimeout(async () => {
+			if (certificate) {
+				await fs.unlinkSync(certificate.path);
+			}
+			if (identity) {
+				await fs.unlinkSync(identity.path);
+			}
+		}, 10000);
 		// TODO: maybe save the data and not the file somewhere in the collection
 		res.status(200).json({ message: 'Successful.' });
 	} catch (err) {
-		console.log(err);
+		console.log('ERROR', err);
 		res
 			.status(400)
 			.json({ error: err.message || 'Could not add new form data.' });
@@ -172,6 +179,69 @@ router.post('/vendor/:id', vendorAuth, async (req, res) => {
 	try {
 		const vendorID = req.params.id;
 		const forms = await Form.find({ vendorID: vendorID }).populate('vendorID');
+		if (!forms) {
+			res
+				.status(400)
+				.json({ error: 'Could not save form data for id:' + vendorID });
+		}
+		res.status(200).json(forms.map((form) => form.toJSON()));
+	} catch (err) {
+		res.status(400).json({ error: 'Could not load the form data.' });
+	}
+});
+
+// @Route   POST api/form/school/add
+// @desc    Add new school form
+// @access  Public
+router.post('/school/add', async (req, res) => {
+	try {
+		const {
+			schoolID,
+			email,
+			fullName,
+			phoneNumber,
+			dateOfVisit,
+			timeOfVisit,
+			studentID,
+			roomNumber,
+		} = req.body;
+
+		const newForm = new SchoolForm({
+			schoolID,
+			email,
+			fullName,
+			phoneNumber,
+			dateOfVisit,
+			timeOfVisit,
+			studentID,
+			roomNumber,
+		});
+		let savedForm = await newForm.save();
+
+		if (!savedForm) {
+			res.status(400).json({ error: 'Could not save form data.' });
+		}
+
+		const savedFormPopulated = await savedForm
+			.populate('schoolID')
+			.execPopulate();
+
+		await awsSES.schoolFormEmail_SES(savedFormPopulated);
+
+		res.status(200).json(savedForm.toJSON());
+	} catch (err) {
+		console.log(err);
+		return res.status(400).json({ error: 'Could not load the form data.' });
+	}
+});
+
+// @Route   POST api/form/school/:schoolid
+// @desc    Get individual form
+// @access  Vendor
+router.post('/school/:id', schoolAuth, async (req, res) => {
+	try {
+		const schoolID = req.params.id;
+		const forms = await SchoolForm.find({ schoolID });
 		if (!forms) {
 			res
 				.status(400)
